implemented the join game button

+++ app/Http/Controllers/EventController.php
@@ -38,6 +38,10 @@ class EventController extends Controller implements MessageComponentInterface
     const CALL_AUTHENTICATE = 'org.cah.authenticate';
     const CALL_PLAYER_INFO = 'org.cah.player.info';
     const CALL_CARDSETS_LIST = 'org.cah.cardset.list';
+
+    //the game the player is currently playing
+    const CALL_GAME_GET = 'org.cah.game.get';
+    const CALL_GAME_JOIN = 'org.cah.game.join';
     const CALL_GAME_LIST = 'org.cah.game.list';
     const CALL_GAME_DELETE = 'org.cah.game.delete';
     const CALL_GAME_CREATE = 'org.cah.game.create';
@@ -141,6 +145,14 @@ class EventController extends Controller implements MessageComponentInterface
             case static::CALL_CARDSETS_LIST:
                 $conn->send($this->encodeMessage($callId, static::RESPONSE_CODE_SUCCESS, 'sucessfull', Cardset::all()));
                 break;
+            case static::CALL_GAME_JOIN:
+                $game = Game::publicId($parameters['gameId'])->first();
+                if ($player->joinGame($game)) {
+                    $conn->send($this->encodeMessage($callId, static::RESPONSE_CODE_SUCCESS, 'successfull', $game));
+                } else {
+                    $conn->send($this->encodeMessage($callId, static::RESPONSE_CODE_SUCCESS, 'already joined', $game));
+                }
+                break;
             case static::CALL_GAME_LIST:
                 if ($player->can('list', Game::class)) {
                     $conn->send($this->encodeMessage($callId, static::RESPONSE_CODE_SUCCESS, 'successfull', Game::all()));
@@ -149,17 +161,25 @@ class EventController extends Controller implements MessageComponentInterface
                 }
                 break;
             case static::CALL_GAME_DELETE:
-                $game = Game::publicId($parameters['gameId'])->first();
-                GameController::deleteGame($game);
-                $conn->send($this->encodeMessage($callId, static::RESPONSE_CODE_SUCCESS, 'successfull'));
+                if ($player->can('destroy', $game)) {
+                    $game = Game::publicId($parameters['gameId'])->first();
+                    GameController::deleteGame($game);
+                    $conn->send($this->encodeMessage($callId, static::RESPONSE_CODE_SUCCESS, 'successfull'));
+                } else {
+                    $this->sendNotAuthorizedResponse($conn, $callId);
+                }
                 break;
             case static::CALL_GAME_CREATE:
-                $game = new Game([
-                    'name' => $player->username . 's Game'
-                ]);
-                $game->public_id = GameController::generateNewPublicId();
-                $game = $player->createGame($game);
-                $conn->send($this->encodeMessage($callId, static::RESPONSE_CODE_SUCCESS, 'successfull'));
+                if ($player->can('store', Game::class)) {
+                    $game = new Game([
+                        'name' => $player->username . 's Game'
+                    ]);
+                    $game->public_id = GameController::generateNewPublicId();
+                    $game = $player->createGame($game);
+                    $conn->send($this->encodeMessage($callId, static::RESPONSE_CODE_SUCCESS, 'successfull'));
+                } else {
+                    $this->sendNotAuthorizedResponse($conn, $callId);
+                }
                 break;
             default:
                 $conn->send($this->encodeMessage($callId, static::RESPONSE_CODE_UNKNOWN_CALL, 'unknown call'));
